feat: Change float to double

+++ cpp-module/09/ex00/BitcoinExchange.cpp
@@ -12,8 +12,8 @@ BitcoinExchange::BitcoinExchange(const std::string& file) {
 	std::string line;
 	std::string date;
 	std::string value;
-  float d_value;
-  std::pair<std::string, float> pair;
+  double d_value;
+  std::pair<std::string, double> pair;
 
   infile.open(file);
   if (infile.is_open() == false) {
@@ -75,9 +75,9 @@ bool BitcoinExchange::isValidHeader(std::ifstream& infile, const std::string& fi
 
 void BitcoinExchange::parseLine(std::string& line) {
   try {
-    std::pair<std::string, float> pair;
-    std::map<std::string, float>::iterator it;
-    float balance;
+    std::pair<std::string, double> pair;
+    std::map<std::string, double>::iterator it;
+    double balance;
 
     pair = splitKeyPair(line, '|');
     it = this->chart_.find(pair.first);
@@ -92,13 +92,13 @@ void BitcoinExchange::parseLine(std::string& line) {
   }
 }
 
-std::pair<std::string, float> BitcoinExchange::splitKeyPair(std::string& str,
+std::pair<std::string, double> BitcoinExchange::splitKeyPair(std::string& str,
                                                              char delimiter) {
   std::string key;
   std::string value;
-  float d_value;
+  double d_value;
   std::stringstream ss(str);
-  std::pair<std::string, float> pair;
+  std::pair<std::string, double> pair;
 
   d_value = 0;
   std::getline(ss, key, delimiter);
